LayoutTests/fast/canvas/canvas-backing-store-reuse-expected.txt

PASS data.data is [0, 0, 0, 0]
PASS successfullyParsed is true

TEST COMPLETE
